requirements.txt
streamlit
scikit-learn
tensorflow
keras
Pillow
st-pages